ipPhone_Project/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
Markdown==3.10
packaging==25.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0